Added more clarifying text to the title as well as implemented a winning scenario
Added an if statement in the play scene so that once all blips are collected the game will finish. Made the title screen make more sense and placed text below each character sprite to make sure the player knows who is who. In the game over scene it determines who is the winner of the match played.